Extract getScopes
